+++ CHANGES.rst
@@ -176,6 +176,8 @@ tweakreg
 - Added the 'GAIADR3' catalog to the available options for alignment; 
   this has been enabled as the default option [#7611].
 
+- GAIADR3 catalog positions are corrrected for proper motions given an epoch [#XXXX]. 
+
 
 1.10.2 (2023-04-14)
 ===================